requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues§

azure-functions
pandas
Prophet
numpy
scikit-learn
sqlalchemy
psycopg2-binary
azure-data-tables>=12.4.0
python-dotenv
openai>=1.0.0  # For LLM integration
aiohttp>=3.8.0  # For async HTTP requests